refactor(image): 게시물 삭제 메서드 내 처리 개선

+++ src/main/java/org/example/story/domain/blog/service/BlogService.java
@@ -96,7 +96,9 @@ public class BlogService {
 
         BlogJpaEntity blog = blogRepository.findByIdAndUserId(blogId, userId)
                 .orElseThrow(() -> new ExpectedException(HttpStatus.NOT_FOUND, "존재하지 않는 블로그입니다."));
-        imageService.deleteImage(blog.getThumbnail());
+        if (blog.getThumbnail() != null) {
+            imageService.deleteImage(blog.getThumbnail());
+        }
         blogRepository.delete(blog);
     }
